[style] change dir
