+++ src/utils/argv_check_utils.c
@@ -1,23 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   return_outfile_fd.c                                :+:      :+:    :+:   */
+/*   argv_check_utils.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yliu <yliu@student.42.jp>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/19 18:04:47 by yliu              #+#    #+#             */
-/*   Updated: 2024/04/24 21:55:50 by yliu             ###   ########.fr       */
+/*   Created: 2024/05/04 09:40:35 by yliu              #+#    #+#             */
+/*   Updated: 2024/05/04 09:40:42 by yliu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "process.h"
+#include "utils.h"
 
-int	return_outfile_fd(const char *filename)
+bool	is_first(int cmd_num)
 {
-	int	fd;
+	return (cmd_num == 0);
+}
 
-	fd = open(filename, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-	if (fd == FAIL)
-		exit(print_error(filename, strerror(errno), 1));
-	return (fd);
+bool	is_last(int cmd_num, int argc)
+{
+	return (cmd_num == argc - 4);
+}
+
+bool	is_middle(int cmd_num, int argc)
+{
+	return (cmd_num > 0 && cmd_num < argc - 4);
 }
